+++ libs/indexPredefined.ml
@@ -18,11 +18,7 @@ open Outcometree
 (* List and doc taken from
    http://caml.inria.fr/pub/docs/manual-ocaml/manual034.html (4.00.1) *)
 
-#if OCAML_VERSION >= (4,08,0)
-  let n s = {printed_name = s}
-#else
-  let n s = s
-#endif
+let n s = {printed_name = s}
 
 let mktype name ?(params=[]) ?(def=Otyp_abstract) doc = {
   path = [];
@@ -37,23 +33,13 @@ let mktype name ?(params=[]) ?(def=Otyp_abstract) doc = {
                               ot_name = v;
                               ot_variance = Asttypes.(NoVariance, NoInjectivity) }
                           ) params;
-  #elif OCAML_VERSION >= (4,12,0)
-        otype_params  = List.map (fun v -> v,Asttypes.(NoVariance, NoInjectivity)) params;
   #else
-        otype_params  = List.map (fun v -> v,(true,true)) params;
+        otype_params  = List.map (fun v -> v,Asttypes.(NoVariance, NoInjectivity)) params;
   #endif
         otype_type    = def;
         otype_private = Asttypes.Public;
-  #if OCAML_VERSION >= (4,03,0)
-    #if OCAML_VERSION >= (4,10,0)
         otype_immediate = Type_immediacy.Unknown;
-    #else
-        otype_immediate = false;
-    #endif
-    #if OCAML_VERSION >= (4,04,0)
         otype_unboxed = false;
-    #endif
-  #endif
         otype_cstrs   = [] }, Orec_not));
 loc_sig = Lazy.from_val Location.none;
   loc_impl = Lazy.from_val Location.none;
@@ -72,16 +58,8 @@ let mkvariant name parent params = {
         otype_type    = (match params with [] -> Otyp_sum []
                                          | l  -> Otyp_tuple l);
         otype_private = Asttypes.Public;
-  #if OCAML_VERSION >= (4,03,0)
-    #if OCAML_VERSION >= (4,10,0)
         otype_immediate = Type_immediacy.Unknown;
-    #else
-        otype_immediate = false ;
-    #endif
-    #if OCAML_VERSION >= (4,04,0)
         otype_unboxed = false;
-    #endif
-  #endif
         otype_cstrs   = [] }, Orec_not));
   loc_sig = Lazy.from_val Location.none;
   loc_impl = Lazy.from_val Location.none;
@@ -124,15 +102,11 @@ let var name = Otyp_var (false, name)
 let constr ?(params=[]) name =
   Otyp_constr (Oide_ident name, List.map var params)
 
-#if OCAML_VERSION >= (4,14,0)
 let out_constr ~name ~args ~ret = {
   Outcometree.ocstr_name = name;
   Outcometree.ocstr_args = args;
   Outcometree.ocstr_return_type = ret;
 }
-#else
-let out_constr ~name ~args ~ret = (name, args, ret)
-#endif
 
 let ibool =
   mktype "bool"